+++ src/Microsoft.VisualStudio.ProjectSystem.IntegrationTests/Environment/ProjectSystemHostConfiguration.cs
